requirements.txt
Flask
Flask-CORS
nltk
cmudict
ffmpeg-python
openai-whisper
difflib
sqlite3
sqlalchemy
werkzeug
montreal-forced-aligner
streamlit